requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==24.3.0
authx==1.4.1
Automat==24.8.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
chromedriver-autoinstaller==0.6.4
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.1
cssselect==1.2.0
defusedxml==0.7.1
dnspython==2.7.0
docutils==0.21.2
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
filelock==3.0.12
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==2.8
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jwt==1.3.1
lockfile==0.12.2
lxml==5.3.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
nltk==3.9.1
numpy==2.2.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsel==1.10.0
passlib==1.7.4
Protego==0.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyDispatcher==2.0.7
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
PySocks==1.7.1
python-daemon==2.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
queuelib==1.7.0
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
Scrapy==2.12.0
selenium==4.29.0
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sql==2022.4.0
starlette==0.41.3
tldextract==5.1.3
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
Twisted==24.11.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
w3lib==2.3.1
watchfiles==1.0.4
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0
wsproto==1.2.0
zope.interface==7.2
